Add integration tests for requests_client
